# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Property api Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://property.test/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/properties": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Récupéré la liste des propriétés",
                "description": "Retourne la liste complète de toute les propriétés.",
                "operationId": "getProperties",
                "responses": {
                    "200": {
                        "description": "Opération réussis",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "400": {
                        "description": "Requête erroné"
                    },
                    "404": {
                        "description": "Page introuvable"
                    }
                }
            }
        },
        "/property/{id}": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Récupéré une propriété",
                "description": "Retourne une propriété.",
                "operationId": "getOneProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "l'id de la propriété",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussis",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "400": {
                        "description": "Requête erroné"
                    },
                    "404": {
                        "description": "Page introuvable"
                    }
                }
            }
        },
        "/property/create": {
            "post": {
                "tags": [
                    "Tests"
                ],
                "summary": "Ajouter une propriété",
                "operationId": "AddOneProperty",
                "requestBody": {
                    "description": "Proprietes",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "image",
                                    "postcode",
                                    "city",
                                    "address",
                                    "room",
                                    "price",
                                    "size",
                                    "floor"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "postcode": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "room": {
                                        "type": "number"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "size": {
                                        "type": "number"
                                    },
                                    "floor": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussis",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "400": {
                        "description": "Requête erroné"
                    },
                    "404": {
                        "description": "Page introuvable"
                    }
                }
            }
        },
        "/property/{id}/update": {
            "put": {
                "tags": [
                    "Tests"
                ],
                "summary": "Modifier une propriété",
                "operationId": "UpdateOneProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "l'id de la propriété",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Proprietes",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "image",
                                    "postcode",
                                    "city",
                                    "address",
                                    "room",
                                    "price",
                                    "size",
                                    "floor"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "postcode": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "room": {
                                        "type": "number"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "size": {
                                        "type": "number"
                                    },
                                    "floor": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussis",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "400": {
                        "description": "Requête erroné"
                    },
                    "404": {
                        "description": "Page introuvable"
                    }
                }
            }
        },
        "/property/{id}/delete": {
            "delete": {
                "tags": [
                    "Tests"
                ],
                "summary": "Delete one porperty",
                "operationId": "deleteOnePorperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussis",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "400": {
                        "description": "Requête erroné"
                    },
                    "404": {
                        "description": "Page introuvable"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}